Split indirect collects into their own test case
